Add special page to import an entity
Special:ImportEntity imports a single entity and then redirects to the
imported entity.

Two new configuration options are added to provide a human-readable link
to the remote Wikibase installation on the special page, in addition to
the API URL that the extension needs to actually import the entity.

Importing statements does not work properly, since the page_props table
is apparently not updated at all anymore once the first attempt to
import statements is made. For now, the checkbox to import statements is
disabled, and statements are never imported.

This commit is not yet ready for merging, since it currently includes
two changes to EntityImporterFactory; the change from private to public
might be necessary, but the commented-out toplevel statements probably
have some better solution.